Alert message if adding duplicate task

+++ js/main.js
@@ -6,10 +6,7 @@ function taskObj (task_id, task_name, date_time) {
 
 var task_array = [];
 
-/*function find_id(finder) {
-	var a = $(finder).attr("id");
-	return a;
-}*/
+//delete task function
 
 function delete_task(find_id) {
 	for (var i=0; i<task_array.length; i++) {
@@ -19,36 +16,65 @@ function delete_task(find_id) {
     }
 }
 
-var newListElement = '<li><a onclick="return remove_item(this)" id="task%item_id%" href="" class="close_box" id="close_box%item_id%">x</a><input type="checkbox" name="task_name" class="strikethrough"><span>%data%</span><span id="task_time%item_id%">%current_time%</span></li>';
+var newListElement = '<li><a onclick="return remove_item(this)" id="task%item_id%" href="" class="close_box" id="close_box%item_id%">x</a><input type="checkbox" name="task_name" class="strikethrough"><span>%data%</span><span class="time_class" id="task_time%item_id%">%current_time%</span></li>';
 //return $(elem).siblings("[id]")
 
 function setFocusToTextBox(){
     document.getElementById("custom_textbox").focus();
 };
 
+//checking if the item already in the list
+
+function contains (value) {
+  var doesContain = false
+
+  for (var i = 0, length = task_array.length; i < length; i++) {
+    if (task_array[i].task_name === value) {
+      doesContain = true
+      break
+    }
+  }
+
+  return doesContain
+}
+
+//adding new item to HTML and JS array
+
 var nextId=1;
 function add_item() {
 	var showText = document.getElementById("custom_textbox");
-	var newListElement_edit = newListElement.replace("%data%", showText.value).replace(/%item_id%/g, nextId).replace("%current_time%", time_getter());
-	var new_task = new taskObj('task'+this.nextId, showText.value, time_getter());
-	task_array.push(new_task); 
-	nextId+=1;
-	$("#list").append(newListElement_edit);
-	$('p').hide();
-	$(".paragraph").css('height', '20px');   	
-	document.getElementById('custom_textbox').value='';
-	setFocusToTextBox();
+	if (contains (showText.value)) {
+		window.alert("This Task is already in the list");
+	} else {
+
+		var newListElement_edit = newListElement.replace("%data%", showText.value).replace(/%item_id%/g, nextId).replace("%current_time%", "Added "+time_getter());
+		var new_task = new taskObj('task'+this.nextId, showText.value, time_getter());
+		task_array.push(new_task); 
+		nextId+=1;
+		$("#list").append(newListElement_edit);
+		$('p').hide();
+		$(".paragraph").css('height', '20px');   	
+		document.getElementById('custom_textbox').value='';
+		setFocusToTextBox();
+	} 
     	
 }
 
+//removing item when clicking x box
+
 function remove_item(close_a) {
+	if (confirm("Are you sure you want to delete this item?")) {
+
 	delete_task($(close_a).attr("id"));
 		
 
 	$(close_a).parent('li').remove();
 	
 	return false;
+	}
 }
+
+//function to add time
 
 function time_getter() {
 	var now = new Date();
@@ -64,6 +90,8 @@ function time_getter() {
 	var vremya = m + "/" + d + "/" + y + " " + e + ":" + f;
 	return vremya;
 }
+
+//add functionality to be able to add item when pressing Enter
 
 $( document ).ready(function() {
 	$("#custom_textbox").keyup(function(event){
